Stash & delay diagnostics for placeholders as bugs in item signatures
Instead of manually keeping track of whether to suppress "bad placeholder"
diagnostics via flag `ItemCtxt.suppress_placeholder_errors` just stash
the relevant diagnostics beforehand & steal them later on to delay them
as bugs.
