+++ compiler/rustc_hir_analysis/src/lib.rs
@@ -92,7 +92,6 @@ use rustc_session::diagnostics::feature_err;
 use rustc_span::{ErrorGuaranteed, Span};
 use rustc_trait_selection::traits;
 
-pub use crate::collect::{suggest_impl_trait, suggest_ret_ty_in_typeck};
 use crate::hir_ty_lowering::HirTyLowerer;
 
 fn check_c_variadic_abi(tcx: TyCtxt<'_>, decl: &hir::FnDecl<'_>, abi: ExternAbi, span: Span) {
