+++ tests/ui/async-await/issues/issue-95307.rs
@@ -6,9 +6,6 @@
 pub trait C {
     async fn new() -> [u8; _];
     //~^ ERROR: the placeholder `_` is not allowed within types on item signatures for opaque types
-    //~| ERROR: the placeholder `_` is not allowed within types on item signatures for opaque types
-    //~| ERROR: the placeholder `_` is not allowed within types on item signatures for opaque types
-    //~| ERROR: the placeholder `_` is not allowed within types on item signatures for opaque types
 }
 
 fn main() {}
