+++ tests/ui/closures/missing-body.stderr
@@ -1,9 +1,3 @@
-error[E0121]: the placeholder `_` is not allowed within types on item signatures for closures
-  --> $DIR/missing-body.rs:5:23
-   |
-LL | fn main() { |b: [str; _]| {}; }
-   |                       ^ not allowed in type signatures
-
 error[E0277]: the size for values of type `str` cannot be known at compilation time
   --> $DIR/missing-body.rs:5:17
    |
@@ -12,6 +6,12 @@ LL | fn main() { |b: [str; _]| {}; }
    |
    = help: the trait `Sized` is not implemented for `str`
    = note: slice and array elements must have `Sized` type
+
+error[E0121]: the placeholder `_` is not allowed within types on item signatures for closures
+  --> $DIR/missing-body.rs:5:23
+   |
+LL | fn main() { |b: [str; _]| {}; }
+   |                       ^ not allowed in type signatures
 
 error: aborting due to 2 previous errors
