+++ tests/ui/delegation/generics/generics-gen-args-errors.stderr
@@ -458,12 +458,6 @@ error: inferred lifetimes are not allowed in delegations as we need to inherit s
 LL |     reuse Trait::<'static, 'static>::foo as bar2;
    |           ^^^^^
 
-error[E0121]: the placeholder `_` is not allowed within types on item signatures for functions
-  --> $DIR/generics-gen-args-errors.rs:103:11
-   |
-LL |     reuse Trait::<'static, 'static>::foo as bar2;
-   |           ^^^^^ not allowed in type signatures
-
 error[E0107]: trait takes 3 lifetime arguments but 0 lifetime arguments were supplied
   --> $DIR/generics-gen-args-errors.rs:107:11
    |
@@ -559,12 +553,6 @@ error: inferred lifetimes are not allowed in delegations as we need to inherit s
    |
 LL |     reuse Trait::<'static>::foo as bar5;
    |           ^^^^^
-
-error[E0121]: the placeholder `_` is not allowed within types on item signatures for functions
-  --> $DIR/generics-gen-args-errors.rs:117:11
-   |
-LL |     reuse Trait::<'static>::foo as bar5;
-   |           ^^^^^ not allowed in type signatures
 
 error[E0107]: trait takes 3 lifetime arguments but 1 lifetime argument was supplied
   --> $DIR/generics-gen-args-errors.rs:122:11
@@ -789,6 +777,18 @@ error[E0747]: constant provided when a type was expected
    |
 LL |     reuse foo::<{}, {}, {}> as bar8;
    |                 ^^
+
+error[E0121]: the placeholder `_` is not allowed within types on item signatures for functions
+  --> $DIR/generics-gen-args-errors.rs:103:11
+   |
+LL |     reuse Trait::<'static, 'static>::foo as bar2;
+   |           ^^^^^ not allowed in type signatures
+
+error[E0121]: the placeholder `_` is not allowed within types on item signatures for functions
+  --> $DIR/generics-gen-args-errors.rs:117:11
+   |
+LL |     reuse Trait::<'static>::foo as bar5;
+   |           ^^^^^ not allowed in type signatures
 
 error: aborting due to 74 previous errors
