+++ tests/ui/macros/issue-118048.rs
@@ -6,6 +6,5 @@ macro_rules! foo {
 
 foo!(_);
 //~^ ERROR the placeholder `_` is not allowed within types on item signatures for functions
-//~| ERROR the placeholder `_` is not allowed within types on item signatures for functions
 
 fn main() {}
