+++ tests/ui/macros/issue-118048.stderr
@@ -4,14 +4,6 @@ error[E0121]: the placeholder `_` is not allowed within types on item signatures
 LL | foo!(_);
    |      ^ not allowed in type signatures
 
-error[E0121]: the placeholder `_` is not allowed within types on item signatures for functions
-  --> $DIR/issue-118048.rs:7:6
-   |
-LL | foo!(_);
-   |      ^ not allowed in type signatures
-   |
-   = note: duplicate diagnostic emitted due to `-Z deduplicate-diagnostics=no`
-
-error: aborting due to 2 previous errors
+error: aborting due to 1 previous error
 
 For more information about this error, try `rustc --explain E0121`.
