+++ tests/ui/parallel-rustc/recursive-impl-trait-deadlock-issue-129912.rs
@@ -14,11 +14,11 @@ fn tuple() -> impl Sized {
     (tuple(),)
 }
 
-fn array() -> _ {
+fn array() -> _ { //~ ERROR the placeholder `_` is not allowed
     [array()]
 }
 
-fn ptr() -> _ {
+fn ptr() -> _ { //~ ERROR the placeholder `_` is not allowed
     &ptr() as *const impl Sized
     //~^ ERROR `impl Trait` is not allowed in cast expression types
 }
@@ -41,7 +41,7 @@ fn closure_ref_capture() -> impl Sized {
     }
 }
 
-fn closure_sig() -> _ {
+fn closure_sig() -> _ { //~ ERROR the placeholder `_` is not allowed
     || closure_sig()
 }
